+++ ethereum/api/src/test/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/EthSubmitWorkTest.java
@@ -103,7 +103,7 @@ public class EthSubmitWorkTest {
     final JsonRpcRequestContext request =
         requestWithParams(
             Bytes.ofUnsignedLong(expectedFirstOutput.getNonce()).trimLeadingZeros().toHexString(),
-            Bytes.wrap(expectedFirstOutput.getPowHash()).toHexString(),
+            expectedFirstOutput.getPowHash().toHexString(),
             expectedFirstOutput.getMixHash().toHexString());
     final JsonRpcResponse expectedResponse =
         new JsonRpcSuccessResponse(request.getRequest().getId(), true);